% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.gapdata.R
\name{as.gapdata}
\alias{as.gapdata}
\title{Make a gapdata class object}
\usage{
as.gapdata(d, segments, labels, ...)
}
\arguments{
\item{d}{dendrogram class object}

\item{segments}{a data.frame containing segments information}

\item{labels}{a data.frame containing labels information}

\item{...}{ignored}
}
\value{
the gapdata class object
}
\description{
This function generates a gapdata class object.  This object is used for drawing dendrograms and heatmaps.
}
\keyword{internal}
