% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_branch_positions.R
\name{assign_branch_positions}
\alias{assign_branch_positions}
\title{Re-evaluate the position of branches}
\usage{
assign_branch_positions(d, verbose = FALSE, ...)
}
\arguments{
\item{d}{dendrogram class object}

\item{verbose}{logical for whether in verbose mode or not}

\item{...}{ignored}
}
\value{
the reevaluated dendrogram object
}
\description{
This function reevaluate the position of branches based on the gaps calculated.
}
\keyword{internal}
