% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_positions.R
\name{assign_positions}
\alias{assign_positions}
\title{Re-evaluate the position of leaves}
\usage{
assign_positions(d, runningX = 1, verbose = FALSE, ...)
}
\arguments{
\item{d}{dendrogram class object}

\item{runningX}{numerical position of leaf node on display}

\item{verbose}{logical for whether in verbose mode or not}

\item{...}{ignored}
}
\value{
the reevaluated dendrogram object
}
\description{
This function reevaluate the position of leaves based on the gaps calculated.
}
\keyword{internal}
