% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_list.R
\name{extract_list}
\alias{extract_list}
\title{Extract a list from the dendrogram object}
\usage{
extract_list(d, type, segments_df = NULL, labels_df = NULL, ...)
}
\arguments{
\item{d}{dendrogram class object}

\item{type}{either "triangular" or "rectangular". It determines the same of branches.}

\item{segments_df}{data.frame storing the segment information}

\item{labels_df}{data.frame storking the label positions}

\item{...}{ignored}
}
\value{
the extracted list
}
\description{
This function extract list of data.frames for drawing dendrograms
}
\keyword{internal}
