% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_number.R
\name{format_number}
\alias{format_number}
\title{Function to format a number}
\usage{
format_number(x)
}
\arguments{
\item{x}{a floating number}
}
\value{
formatted number
}
\description{
This function takes a floating number and round to 2 decimal point
}
\keyword{internal}
