% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap_dendrogram.R
\name{gap_dendrogram}
\alias{gap_dendrogram}
\title{Function to draw a gapped dendrogram}
\usage{
gap_dendrogram(
  data,
  leaf_labels = TRUE,
  rotate_label = FALSE,
  orientation = c("top", "right", "bottom", "left"),
  ...
)
}
\arguments{
\item{data}{gapdata class object}

\item{leaf_labels}{a logical to show labels or not}

\item{rotate_label}{a logical to rotate labels or not}

\item{orientation}{a character to set the orientation of dendrogram. Choices are "top", "right", "bottom", "left".}

\item{...}{ignored}
}
\value{
a ggplot object
}
\description{
This function draws a gapped dendrogram using the ggplot2 package. The input for the
function is the gapdata class object, generated from gap_data() function.
}
