% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap_heatmap.R
\name{gap_heatmap}
\alias{gap_heatmap}
\title{Function to draw a gapped heatmap}
\usage{
gap_heatmap(
  m,
  row_gap = NULL,
  col_gap = NULL,
  row_labels = TRUE,
  col_labels = TRUE,
  rotate = FALSE,
  col = c("#053061", "#2166AC", "#4393C3", "#92C5DE", "#D1E5F0", "#F7F7F7", "#FDDBC7",
    "#F4A582", "#D6604D", "#B2182B", "#67001F")
)
}
\arguments{
\item{m}{data matrix}

\item{row_gap}{a gapdata class object for rows}

\item{col_gap}{a gapdata class object for columns}

\item{row_labels}{a logical to show labels for rows}

\item{col_labels}{a logical to show lables for columns}

\item{rotate}{a logical to rotate row labels}

\item{col}{colors used for heatmap}
}
\value{
a ggplot object
}
\description{
This function draws a gapped heatmap using the ggplot2 package. The input for the
function are the gapdata class objects, generated from gap_data() function, and the data matrix.
}
