% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_segment_df.R
\name{get_segment_df}
\alias{get_segment_df}
\title{Make a data.frame object}
\usage{
get_segment_df(x0, y0, x1, y1)
}
\arguments{
\item{x0}{x coordinate of point 1}

\item{y0}{y coordinate of point 1}

\item{x1}{x coordinate of point 2}

\item{y1}{y coordinate of point 2}
}
\value{
A data.frame
}
\description{
This function just make a data.frame based on 4 input parameters
}
\keyword{internal}
