% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.gapdata.R
\name{is.gapdata}
\alias{is.gapdata}
\title{Function to check if a object is a gapdata class object}
\usage{
is.gapdata(x)
}
\arguments{
\item{x}{a object}
}
\value{
a logical TRUE or FALSE
}
\description{
This function checks if a object is a gapdata class object.
}
\keyword{internal}
