% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.exp.R
\name{map.exp}
\alias{map.exp}
\title{A function to map values in a range exponentially}
\usage{
map.exp(value, start1, stop1, start2, stop2, scale = 0.5)
}
\arguments{
\item{value}{input value}

\item{start1}{lower bound of the value's current range}

\item{stop1}{upper bound of the value's current range}

\item{start2}{lower bound of the value's taget range}

\item{stop2}{upper bound of the value's target range}

\item{scale}{scale log base}
}
\value{
a numeric value
}
\description{
This function maps a value in one range to another range exponentially.
}
\keyword{internal}
