% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\title{A function to map values in a range}
\usage{
map(value, start1, stop1, start2, stop2)
}
\arguments{
\item{value}{input value}

\item{start1}{lower bound of the value's current range}

\item{stop1}{upper bound of the value's current range}

\item{start2}{lower bound of the value's taget range}

\item{stop2}{upper bound of the value's target range}
}
\value{
a numeric value
}
\description{
This function maps a value in one range to another range.
}
\keyword{internal}
