% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_most_right_leaf.R
\name{set_most_right_leaf}
\alias{set_most_right_leaf}
\title{Set the most right leaf object from a dendrogram}
\usage{
set_most_right_leaf(d, d2, ...)
}
\arguments{
\item{d}{dendrogram class object, subtree object}

\item{d2}{dendrogram class object, a leaf to replace the most right}
}
\value{
the dendrogram class object where the most right leaf is replaced
}
\description{
This function replace the most right leaf with provided dendrogram
}
\keyword{internal}
