% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_gap.R
\name{count_gap}
\alias{count_gap}
\title{Count the number of gaps based on a threshold}
\usage{
count_gap(d = d, count = 0, threshold = threshold)
}
\arguments{
\item{d}{dendrogram class object}

\item{count}{count}

\item{threshold}{a numeric value for threshold}
}
\value{
the count of gaps
}
\description{
This function counts the number of gaps based on the cutting the tree method.
It counts the number of branches that are above the threshold distance.
}
\keyword{internal}
