% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap_label.R
\name{gap_label}
\alias{gap_label}
\title{Function to draw a gapped labels}
\usage{
gap_label(data, orientation, label_size = 5)
}
\arguments{
\item{data}{gapdata class object}

\item{orientation}{orientation of the labels, "left", "top", "right", or "bottom"}

\item{label_size}{a numeric to set the label text size}
}
\value{
a ggplot object
}
\description{
This function draws a gapped labels using the ggplot2 package. The input for the
function is the gapdata class object, generated from gap_data() function.
}
