% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_most_left_leaf.R
\name{get_most_left_leaf}
\alias{get_most_left_leaf}
\title{Get the most left leaf object from a dendrogram}
\usage{
get_most_left_leaf(d)
}
\arguments{
\item{d}{dendrogram class object}
}
\value{
the most left leaf object
}
\description{
This function returns the most left leaf object.
}
\keyword{internal}
