% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_most_right_leaf}
\alias{get_most_right_leaf}
\title{Get the most right leaf object from a dendrogram}
\usage{
get_most_right_leaf(d)
}
\arguments{
\item{d}{dendrogram class object}
}
\value{
the most right leaf object
}
\description{
This function returns the most right leaf object.
}
\keyword{internal}

